// REST_API_Iteration_1/src/models/request.bodies/vendor.request.bodies/create.vendor.product.request.body/create.vendor.product.request.body.ts
import { AddVendorProductCategoryRequestBody } from "./add.vendor.product.category.request.body";

export interface CreateVendorProductRequestBody{
    vendorId: number;
    unitPriceEuro: number;
    inventoryLevel: number;
    name: string;
    description: string;
    categories: AddVendorProductCategoryRequestBody[];
}